Rework the epitope mapping
We now need to split epitope info databases by protein to use this tool, but it
is more straightforward to do that than string matching against nominal
protein name strings, which vary based on the polyprotein/subprotein name and
are not used consistently between research papers.

This change also changes the way we search for job directories to speed up
processing, and disables displacement calculations for now. We will re-enable
these with a flag.